# apps/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go + Python 퀀트 플랫폼 API 서버입니다.",
        "title": "Quant Trading API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/stocks/list": {
            "get": {
                "description": "DB에 저장된 중복 없는 종목 심볼 목록을 조회합니다. 검색창의 자동완성(Autocomplete) 기능에 사용됩니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "종목 리스트 조회 (자동완성용)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api_stocks.StockItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stocks/search": {
            "get": {
                "description": "symbol 또는 회사명(영문·한국어)에 q가 포함된 종목을 랭킹 순으로 반환합니다. 순위: 완전일치 \u003e symbol 접두사 \u003e name 접두사 \u003e symbol 부분일치 \u003e name 부분일치.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "종목 자동완성 검색",
                "parameters": [
                    {
                        "type": "string",
                        "description": "검색어 (symbol 또는 회사명 일부)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "최대 반환 개수 (기본 20, 최대 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api_stocks.StockSearchItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stocks/{symbol}/history": {
            "get": {
                "description": "특정 심볼의 회사명과 과거 시세 데이터를 조회합니다. (없으면 자동 수집)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "주식 히스토리 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock Symbol (e.g., NVDA)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quant-server_internal_model.StockHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_api_stocks.StockItem": {
            "type": "object",
            "properties": {
                "symbol": {
                    "type": "string",
                    "example": "NVDA"
                }
            }
        },
        "internal_api_stocks.StockSearchItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Rocket Lab Corporation"
                },
                "symbol": {
                    "type": "string",
                    "example": "RKLB"
                }
            }
        },
...
        "quant-server_internal_model.MarketData": {
            "description": "개별 일자의 시가, 고가, 저가, 종가 및 거래량 정보",
            "type": "object",
            "properties": {
                "bb_l": {
                    "type": "number",
                    "example": 139.8
                },
                "bb_m": {
                    "type": "number",
                    "example": 150
                },
                "bb_u": {
                    "type": "number",
                    "example": 160.2
                },
                "close": {
                    "type": "number",
                    "example": 153.4
                },
                "high": {
                    "type": "number",
                    "example": 155.8
                },
                "low": {
                    "type": "number",
                    "example": 149.1
                },
                "macd": {
                    "type": "number",
                    "example": 1.25
                },
                "macd_h": {
                    "type": "number",
                    "example": 0.45
                },
                "open": {
                    "type": "number",
                    "example": 150.25
                },
                "rsi": {
                    "description": "보조지표 필드",
                    "type": "number",
                    "example": 55.42
                },
                "time": {
                    "type": "string",
                    "example": "2026-02-01"
                },
                "volume": {
                    "type": "number",
                    "example": 1200500
                }
            }
        },
...
        "quant-server_internal_model.StockHistoryResponse": {
            "description": "주식의 기본 정보와 과거 시세 데이터를 포함합니다.",
            "type": "object",
            "properties": {
                "symbol": {
                    "type": "string",
                    "x-order": "1",
                    "example": "NVDA"
                },
                "company_name": {
                    "type": "string",
                    "x-order": "2",
                    "example": "NVIDIA Corporation"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quant-server_internal_model.MarketData"
                    },
                    "x-order": "3"
                }
            }
        }
    }
}